chore: move signature code from libstore to libutil
As it makes more sense to make it available to more consumers.

In this commit, we create a new util: `randomSha1` because `libstore`
depended on sodium to do that in the past. We moved the sodium
initialization and dependency to `libutil`.

Moreover, we keep `getDefaultPublicKeys` inside `libstore` as it relies
on the global variable `settings` to derive this information and is
really only but a consumer of the `libutil`'s signature code.
