+++ src/libstore/path.cc
@@ -1,6 +1,6 @@
 #include "store-dir-config.hh"
 
-#include <sodium.h>
+#include "util.hh"
 
 namespace nix {
 
@@ -49,9 +49,7 @@ StorePath StorePath::dummy("ffffffffffffffffffffffffffffffff-x");
 
 StorePath StorePath::random(std::string_view name)
 {
-    Hash hash(HashAlgorithm::SHA1);
-    randombytes_buf(hash.hash, hash.hashSize);
-    return StorePath(hash, name);
+    return StorePath(Hash::random(HashAlgorithm::SHA1), name);
 }
 
 StorePath StoreDirConfig::parseStorePath(std::string_view path) const
