+++ src/nix/verify.cc
@@ -5,6 +5,7 @@
 #include "thread-pool.hh"
 #include "references.hh"
 #include "signals.hh"
+#include "libstore/keys.hh"
 
 #include <atomic>
